+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### 2.x.x
 * Fixed the issue #6
 * Changed the function name of `selectRange` to `selectNextElement`
+* Added new option `cleanValuesOnRestart`
+* Fixed the issue #9
 
 ### 2.0.16
 * Deleted a variable never used